modified for permit the xcode compilation(this script is called also by xcode before compilation)

+++ chaos/common/version.sh
@@ -3,16 +3,20 @@
 BASEDIR=$(dirname $0)
 
 
-export PATH=$PATH:/opt/local/bin:/usr/local/bin
+#export PATH=$PATH:/opt/local/bin:/usr/local/bin
 
-git log -n 1 --pretty="format:#define CSLIB_VERSION_HEADER \"!CHAOS Library Developed By Claudio Bisegni\"%n#define CSLIB_VERSION_NUMBER \"Version:%h\"%n" > /tmp/version.h
-stty -echo
+if [ ! -n $XCODE_PRODUCT_BUILD_VERSION ]; then
+	git log -n 1 --pretty="format:#define CSLIB_VERSION_HEADER \"!CHAOS Library Developed By Claudio Bisegni\"%n#define CSLIB_VERSION_NUMBER \"Version:%h\"%n" > /tmp/version.h
+	stty -echo
 
-if ! diff /tmp/version.h $BASEDIR/version.h>/dev/null; then
-    stty echo
-    echo "Writing version header file" $BASEDIR
-    cp /tmp/version.h $BASEDIR/version.h
+	if ! diff /tmp/version.h $BASEDIR/version.h>/dev/null; then
+		stty echo
+		echo "Writing version header file" $BASEDIR
+		cp /tmp/version.h $BASEDIR/version.h
+	fi
+	rm /tmp/version.h
+	stty echo
+else
+	git log -n 1 --pretty="format:#define CSLIB_VERSION_HEADER \"!CHAOS Library Developed By Claudio Bisegni\"%n#define CSLIB_VERSION_NUMBER \"Version:%h\"%n" > $BASEDIR/version.h
+
 fi
-rm /tmp/version.h
-
-stty echo
